requirements.txt
faker
psycopg2-binary
asyncpg
fastapi
uvicorn
pydantic